Fix and refactor type hint
The previous approach to @option_group didn't work.
When combined with a constraint used as decorator,
mypy couldn't infer a type variable.
The solution was simple: just declare param decorators
without generics. A precise inference is not much
useful here as it doesn't contribute to "propagate"
the type of the command function.

Useful types were moved to typing.py.